Jogosultságok ellenőrzése, Hibák dobása, Eljárások módosítása
createCompany: útólagos cég létrehozásnál a felhasználóhoz nem lett beírva a létrehozott cég id (javítva)

+++ Backend/SzakemberKereso/src/main/java/szakemberkereso/Service/CompaniesService.java
@@ -4,14 +4,18 @@
  */
 package szakemberkereso.Service;
 
+import java.util.Objects;
+import javax.mail.AuthenticationFailedException;
+import javax.ws.rs.ForbiddenException;
+import szakemberkereso.Configuration.Roles;
 import szakemberkereso.Model.Companies;
+import szakemberkereso.Model.Users;
 
 /**
  *
  * @author Sharkz
  */
 public class CompaniesService {
-    
     
     public Companies getCompanyById(Integer id) throws Exception{
         Companies result = Companies.getCompanyById(id);
@@ -30,15 +34,41 @@ public class CompaniesService {
             company.getAddress().setDoor(-1);
         }
         
-        Companies.createCompany(company);
+        //USER jogosultsággal nem hozhat létre céget, és csak magának adhatja hozzá (currentUserId)
+        if (AuthService.isUserAuthorized(company.getCurrentUserId(), new Roles[]{Roles.ADMIN, Roles.WORKER})){
+            Companies.createCompany(company);
+        }
+        else{
+            throw new ForbiddenException("Nincs jogosultsága ehhez a kéréshez.");
+        }
     }
     
     public void updateCompanyById(Companies company) throws Exception{
-        Companies.updateCompanyById(company);
+        //USER jogosultsággal nem lehet cége
+        if (AuthService.isUserAuthorized(company.getCurrentUserId(), new Roles[]{Roles.ADMIN, Roles.WORKER})){
+            //a user-ek csak a saját cégeiket módosíthatják
+            if(!Objects.equals(company.getId(), Users.getUserById(company.getCurrentUserId()).getCompanyId())){
+                throw new ForbiddenException("Nincs jogosultsága ehhez a kéréshez.");
+            }
+            Companies.updateCompanyById(company);
+        }
+        else{
+            throw new ForbiddenException("Nincs jogosultsága ehhez a kéréshez.");
+        }
     }
     
     public void deleteCompanyById(Companies company) throws Exception{
-        Companies.deleteCompanyById(company);
+        //USER jogosultsággal nem lehet cége
+        if (AuthService.isUserAuthorized(company.getCurrentUserId(), new Roles[]{Roles.ADMIN, Roles.WORKER})){
+            //a user-ek csak a saját cégeiket törölhetik
+            if(!Objects.equals(company.getId(), Users.getUserById(company.getCurrentUserId()).getCompanyId())){
+                throw new ForbiddenException("Nincs jogosultsága ehhez a kéréshez.");
+            }
+            Companies.deleteCompanyById(company);
+        }
+        else{
+            throw new ForbiddenException("Nincs jogosultsága ehhez a kéréshez.");
+        }
     }
     
 }
